When plot = FALSE, don't call plot.new
plot.new() is generally unfriendly to call unless it will actually be
used for drawing, and certainly not if we ask for no plots to be
created.

plot.new triggers the completion of the previous plotting action, and as
such can't be used in combination with any multithreading or forked R
processes.

plot.new also does not return a "plot" object, it always returns `NULL`,
so although it appears to be used here as a way of returning a "null
plot", all that really happens is that AnomalyDetection* returns
`list(..., plot = NULL)`.

This commit resolves #60