Add period values for 'sec' and 'ms' granularities, as it appears they are missing.
If the granularity detection results in "sec" or "ms" granularity
detect_anons will blow up with a message stating 'period' was not
provided. Which is correct. This provides at least some defaults.

This is a bit tricky, since the number of samples really isn't guaranteed
to to be 1 measurement per whatever the 'gran' variable says it is.
However, this code appears to assume that. Leading to "fun" when the
precision is in ms but the measurements fire a lot less often - say, once
per five minutes (that's the testcase in question).

There are two ways to deal with this, neither of them implemented by this patch:
- Look at the delta between the first and the last timestamp and the number
of records and make an educated guess from there.
- Add 'period' as a function argument and let the caller decide, using the
hardcoded values simply as educated guesses.

These approaches can be combined, but that is left as an exercise for the reader.